requirements.txt
python-dotenv
amino.fix